3dReHo: store the sorted time series as float, not int
CalcRanksForReHo declared its 'sorted' working array as int, so each
float time-series value was truncated toward zero before the tie test.
Any two distinct values sharing an integer part were then treated as
tied: their ranks were replaced by an average rank and the Kendall's W
tie-correction term was inflated. The effect scales with signal
amplitude, so it biases ReHo differently across datasets -- and on
data of order 1 (percent signal change, unit-variance scaling) every
value collapses to a single tie group.

THD_get_voxel() already returns float, so storing the value as float
makes the comparison lossless.

The separate tie-at-end-of-series problem this branch originally also
addressed is now fixed on master (29384a2, from Kevin Tran), so only
the truncation fix remains here.

+++ src/ptaylor/rsfc.c
@@ -60,7 +60,7 @@ int CalcRanksForReHo(float *IND, int idx, THD_3dim_dataset *T, int *NTIE,
   int LENTIE = 0;
   float TIERANK;
   int *toP=NULL; // to reset permuts
-  int *sorted=NULL; // hold sorted time course, assume has been turned into int
+  float *sorted=NULL; // hold sorted time course
   int val;
 
   // GSL stuff
@@ -69,7 +69,7 @@ int CalcRanksForReHo(float *IND, int idx, THD_3dim_dataset *T, int *NTIE,
 
 
   toP = (int *)calloc(TDIM,sizeof(int)); 
-  sorted = (int *)calloc(TDIM,sizeof(int)); 
+  sorted = (float *)calloc(TDIM,sizeof(float)); 
 
   if( (toP ==NULL) || (sorted ==NULL) ) { 
     fprintf(stderr, "\n\n MemAlloc failure.\n\n");
